+++ integration-tests/src/test/java/org/realityforge/arez/integration/AbstractArezIntegrationTest.java
@@ -21,7 +21,7 @@ public abstract class AbstractArezIntegrationTest
     ArezTestUtil.setEnforceTransactionType( true );
     ArezTestUtil.setEnableSpy( true );
     ArezTestUtil.setEnableZones( false );
-    ArezTestUtil.setProvider( null );
+    ArezTestUtil.clearProvider();
   }
 
   @AfterMethod